SkinSyncApp-ML/requirements.txt
Flask
tensorflow
Pillow
numpy
PyJWT
Werkzeug
requests
Flask-Cors